make `wp_template` not `public` by default
The `public` flag when registering a post type opens it up to all kinds of exposure. Since we only have this in place to support searching through the core endpoint, this simply makes the CPT `public` when in an API endpoint.

We could probably do more robust detection than this to more strictly limit exposure (eg hooking into `WP_Query`) but since it's unlikely that we will continue to use the core search endpoint, this will do fine for now.

Fixes #32491