Remove extensions loader from server config
In #27415 we removed the extensions loader from Calypso.  We didn't
remove the configuration for it from the `webpack.config.node.js` file.

It probably didn't break becuase `extensions/index` wasn't being loaded
on the server and so the rule didn't match and the loader wasn't
requested.

This patch removes the reference that should have disappered in that PR.