Skip the localStorage polyfill in evergreen build. (#32200)
All the browsers supported in the evergreen build include support for
localStorage. This includes Safari, which no longer throws errors when
using localStorage in private windows.

As such, we can replace the polyfill with a noop in the evergreen build.

This change also adds a generic mechanism to skip other (non-coreJS)
polyfills in the evergreen bundle.