Popover + Search Filters Dropdown: Improving focus interaction + accessibility
This update enhances the base `Popover` component with improved keyboard
interactions + accessibility support.

A new Popover prop has been added - `isFocusOnShow` (`bool`). By default,
it is `false`, which ensures that the experience for current Popover use
remains the same.

However, once that prop is set to true, on show, the focus will be shifted
to the `Popover`.

Aria roles, labels, and descriptions have been added to ensure proper
screen-reader/voiceover support.

The first place that the Popover will become focused on open would be within
the Domain Search Filters Dropdown.

Aria labels and ids have been added to the trigger (context) button to
associate it with the Popover.