WIP: o2-blocks: Add KLIPSE code evaluator block
See viebel/klipse

Adds a new block to our internal set of blocks so we can experiment with
interactive chunks of code both on the edit side and also on the view
side. If merged this should open up some new possibilities for building
educational blog posts about code, for demonstrating a specific
technique, or for any number of things where visualizing or manipulating
the results of some code enhances the reading experience.