Simplify a couple of existing types
There's an oddity here in that I encoded a type error:
returning `void` from `unsubscribe` when it actually
returns `bool` or `ReturnType<typeof listeners.delete>`

The reason I did this was actually to enforce the idea
that consumers should not rely on the return value at
all. There is no intention of providing a guarantee on
what that value is. We _could_ bring the code into
harmony with this by turning the function into a non-value
returning function but I don't see much value in doing
that just for the sake of appeasing the types.

At this point consumers should get an error if they use
the return value.

```ts
// technically more correct
return () => { listeners.delete( f ); return }
```