Update Inline Imports: Simplify and remove recursion
The Babel plugin API provides information on existing bindings for
each identifier in the visitor's scope. We can use this information
to find the original import declaration instead of starting at the
top and recursively finding the imported identifiers.

This should speed up the transform and also prevent issues from
having the TypeScript transform apply after this one.