transform-wpcalypso-async Babel plugin: remove test dependency on plugin-babel-syntax-jsx
All the Babel syntax plugins are just convenient helpers to add options to the Babylon parser.
The `@babel/syntax-jsx` plugin doesn't actually contain the JSX parsing code. It just enables
an option on the monolithic parser.

When testing the `transform-wpcalypso-async`, we can eliminate a dependency by passing the option
directly in `parserOpts`.