Remove `inline-imports` Babel Transform
See https://calypsop2.wordpress.com/2017/12/28/action-type-inlining/

We previously included this transform in order to replace imported
constants for our Redux action types with their string-value
equivalents. At the time this saved us around seven to ten kilobytes of
compressed build being sent to our customers.

A bug in the implementation of #33149 suggested however that this
transform may no longer be providing the value it formerly did.

In this PR I'm removing the transform altogether and will see how that
changes the output build size of our bundles. If we can eliminate this
then it will further simplify our build pipeline.