Remove TMP & TODO comments
I had thought that the query parsing belongs in the routing logic, but I believe I was wrong.

The existing page middleware does parse the search string as well as determine the appropriate locale information, but it stores it immediately in the redux store and persists it indefinitely, which is appropriate - a language choice is relevant until a new one is made.

The intention here is much more temporary - we expect these strings to
be loaded up for specific short-term purposes, like seeing how new strings look before being added to the core set and deployed normally.

It is potentially a large amount of additional data, so if we decide to start using this to use this for longer term purposes we should add some caching for the data and reconsider adding the parameters it to the redux tree alongside the current locale data.