Checklist: Infer loading state from getRequest().isLoading (#33307)
Change the implementation of `getSiteChecklistIsLoading` to use `getRequest( state, requestSiteChecklist( siteId ) ).isLoading`, which, thanks to the data layer, gives us the state of the network request linked to `requestSiteChecklist` for free. This allows us to remove the `isLoading` reducer.

Furthermore, rename the `getSiteChecklistIsLoading` selector to `isSiteChecklistLoading` to align with our established pattern.