Remove legacy CSS hot reload code (build on every request)
Express.js middleware that schedules CSS rebuild on every request. Active when
in development mode and having `css-hot-reload` feature flag disabled, which
means never.

The hot reload is always done by the webpack HMR plugin that watches CSS files for
changes and pushes the rebuilt assets to the client through the HRM websocket.
That will be deleted soon, too, after completely migrating to webpack and `style-loader`,
but this is a nice incremental step.