Change first letter in legends
Changing legend to capitalize first letter of all notifications. It balances the sections out better by not having all lowercase, just looks better.

Props to @dreiris for help on working through this as we try to figure out how bottles work!

Ran into three pre-commit hooks that we're going to punt, to be addressed later.