Replace createReducer with keyedReducer for signup verticals (#33802)
This change means that when the reducer is called with `undefined` passed in as the state, the result will be an empty object, instead of `null`. This doesn't seem to have a negative consequence, and the `getVerticals` selector already handles an empty object as `null`. Other reducers in signup also use `createReducer` so they could also be candidates for moving to `keyedReducer` instead.