Removing logic that dumps customers to WP admin
Currently there's logic that will send non US/Canada business plan customers directly to the WP-admin when they begin onboarding. They're sent without a warning that they'll be in a new context.

However, currently US/Canada customres get an appropriate notice letting them know what's about to happen.

This fix removes the logic and makes it so that all customers (regardless of location) are presented with the notice and told about the switch the WP-admin.

In order to pass pre-commit hook I needed to comment out other lines of code as well. I can't tell if they are causing conflicts elsewhere so I absolutely would welcome a review on this. Thank you!