Update tooltip to popover on stats page
URL: https://wordpress.com/stats/day/testingground728.blog

On the stats page of Calypso we have a tooltip that uses a hover instead of select action.

I saw that another page uses a different method that seems to be a better pattern. My guess is this was created later on.

Comparison popover can be seen on: https://wordpress.com/settings/general/testingground728.blog

Updated CSS and code to bring the new popover design into the stats page.