Remove `require-await` eslint rule.
This rule was originally added in #30697 to help catch incorrect usage of
async functions. It validates that an async function has at least one
usage of `await` in its body.

While this helps catch basic errors such as forgetting to remove the
`async` keyword when refactoring an async function to sync, it also
prevents some useful, terse syntax.