Resolving a few warnings/errors
1. selectedSiteId is now no longer a required propType, given that the quickstart offer page can be access while logged out 2. receiptId is no longer ju
t a number in the quickstart offer page 3. If the receiptId val is `add`, then skip the UI and directly take to checkout. This needs to be handled in th
e component life cycle methods, instead of in render() 4. Move redirectLoggedOut to mapStateToProps to avoid action undefined error in wpcom-api-middleware.js