Signup: reset signup state before navigation to destination outside Calypso
The code path that handled navigation outside Calypso (typically site frontend by virtue of
the `getSiteDestination` function) didn't call `signupFlowController.reset()`, leaving the
signup state in Redux (and in IndexedDB) and causing weird behavior should we return to `/start`
again shortly after.

This patch also unifies the code paths that handle inside- and outside-Calypso destinations,
as there is no longer any difference between them. The inside- one used to use `page( ... )`
to navigate, but that's no longer true.

Also removes the `defer` call that was needed for Flux, but is not needed in Redux-only code.