Correct DOM order for back button in signup flow
Reordering the Back button in `<StepWrapper>` so that you shift-tab to
it from the onboarding controls, which is what you'd expect from the
visual order.

The back button can be safetly moved out of the `step-wrapper__button`
because that container is only doing centering and the back button is
absolutely positioned.