Signup: remove code that resets signup state after user is logged in
When mounting the `Signup` UI, there is a check whether the user is logged in and yet
there are data from the `user` step in the progress store. That condition can never
happen in the Reduxified Signup, because logged-out and logged-in Redux state is stored
in different IndexedDB keys. Logged-in session will never read the persisted signup state
that the logged-out session has written.

Nice incremental step to simplify the very complex Signup controller code.