Update {{CompanyName}} placeholders when user types
The site title (company name) appears in the site preview in the contact
information and the copyright notice. This wasn't being updated because
sCU bailed too soon before everything that was affected by the site
title had been updated.

I also updated the check that re-renders the whole iframe while I was
there. A change to the `isRtl` prop now refreshes the entire preview.
Seems like it should be treated the same way as `laungSlug`.