Fix wp-admin link appearing below the fold on Chrome
The .sidebar__menu-wrapper (which contains the wp-admin link)
was 100% of its parent height. However its parent also contains the
current site info, so the bottom of .sidebar__menu-wrapper was pushed
off screen.

Using flex so that we use up all remaining space in the parent, not 100%
of it.

Issue wasn't present on Firefox, but I think using `flex: 1` makes more
sense here IMO so it doesn't feel like a hack to make Chrome work.