Hot: Speed up the hot reloader by memoizing webpack stats
Profiling found that emitting the JSON webpack stats consumed a substantial amount of time for each new connection to the dev server. While this is mostly inconsequential for local development, when running the e2e tests against a local instance it would cause the node process to use up a substantial amount of CPU time.

Memoizing the toJson call substantially reduces the CPU time required when many clients are connecting and disconnecting from the dev server.