Hide blog preview in all signup flows
Creates explicit "no preview" versions of each flow where it's possible
to choose the blog segment. This replaces some intricate branching in
`<SiteType>` and instead moves the user to a different flow depending on
their segment choice.

Another approach might have been to remove the `<SiteMockup>` when
`site-type` was `'blog'`, but then we would still need a way to skip the
`site-style` step (which doesn't make sense if there's no preview), and
we wouldn't be able to have a flow like `onboarding-dev` that allows us
to test the blog preview before shipping it.