Add `once` method to memoizeLast.
This is effectively an alias of `memoizeLast`, albeit with stricter
typing, which ensures that the function to be wrapped doesn't have
any arguments. There's a runtime check too, for non-typed scenarios.

This guarantees that the provided method will only be run once (through
the wrapper), and we can thus take advantage of the special semantics
with some clearer naming.