Update signup canResume's dependencies to only match against steps from the same flow as the current flowName
This attempts to make `resumeProgress` less eager to resume the progress. In this change, progress should only resume if we're certain that the user is resuming the same flow that they were last on. Otherwise, they will start from the beginning of the flow again.

This is a WIP commit to try out an approach for dealing with issue #34299