Use `find_available_url` param if user skips domain step
Being careful to only use the find_available_url param if the user is in
the test group. There might be another reason `siteUrl` could be empty
that I don't want to interfere with.

Was going to add `isdomainStepSkippable` to `signup/utils.js`, but this
created a webpack issue, I'm guessing a circular reference.