Update signup canResume's dependencies to only match steps against the same flowName
This is the second attempt at making `resumeProgress` less eager to resume the progress. In this change, progress should only resume if we're certain that the user is resuming the same flow that they were last on. Otherwise, they will start from the beginning of the flow.

Additionally, `getCompletedSteps`, which gets called by the main Signup's `isEveryStepSubmitted` method should _not_ only match against the same flow. Tests are included to confirm this, and fixes a regression in #34356 where the loading screen would not display after submitting the final step, and instead it would show `user` step before redirecting to the checkout page. That issue was logged in #34403.