Force GdprBanner not to render in SSR.
GdprBanner tries to access certain client-side-only properties, such as
the `cookies` object in document, and navigator.userAgent.

This caused SSR to fail when the component was added to the logged-out
layout.

Since it doesn't make sense to render the GDPR banner on the server
anyway (since we don't have enough information to determine whether it
should be shown), this change effectively disables it whenever the
`document` object isn't present.