Add check for `devdocs/redirect-loggedout-homepage` before redirecting to login page
This change adds a check for 'devdocs/redirect-loggedout-homepage' before updating each section to use the `redirectLoggedOut` controller. In environments like `development` where `devdocs` is enabled, the `loggedOutMiddleware` should redirect requests to `/` to `/devdocs/start` so that new developers see a friendly landing page, and are able to walk through the simpler `/start/developer` signup flow. The existing conditional was essentially overriding this and redirecting `/` to `/log-in` instead.