Color Schemes: Remove duplicated Thumbnails from public/ (#34480)
`public/` is now disposable. Our static assets live in `static/` now, and are copied to `public/` as part of the build process.

It seems like these two SVGs accidentally slipped in with #32794 and #33136, respectively. The same two files were also added to the correct location in `static/`, so we don't have to move them there but can simply drop them.