Fix InfoPopover event propagation
In https://github.com/Automattic/wp-calypso/pull/32805 the handleClick method for the InfoPopover has changed and it seems that broke how the component works inside clickable content. It seems we need to also stopPropagation for the event because otherwise you can't click on the InfoPopover component if it is inside another component that has onClick event handler (for example in domain search we have an HSTS InfoPopover icon and it can't be clicked without adding the domain in your cart). This fixes that.