Action Panel: Allow figure to be left aligned
As part of the project to overhaul the 'Earn' section, we need a card
format that's similar to the action panel, but has the image on the
lefthand side.

This change adds a prop to the figure component, that can be set to
`left` or `right`, with a default of right to be consistent with the
current behaviour.