Components: Add a general promo card component
As part of the project to overhaul the 'Earn' section, we need a card
component that is similar to the card used in the Marketing Tools
section.

This change creates a generalised version of the card component that can
be used in both sections, along with any future sections that follow a
similar pattern.