CreditCardPaymentBox: Only look for errors on current step in isFormSubmitting
When looking for errors in `isFormSubmitting()` for the steps
`RECEIVED_PAYMENT_KEY_RESPONSE` and `RECEIVED_AUTHORIZATION_RESPONSE`,
we check the _current_ props of the component, even though sometimes the
function is checking the previous props or the next props. I'm not
certain, but I think that this is actually just a typo and we should be
always only looking for errors on the current step itself.