Remove redundant hack used to fix some flows
The `premium`, `business`, and `personal` flows were broken when
choosing the blog site type would switch the user to the
`onboarding-blog` flow. This carefully crafted hack prevented the flow
from being switched when the user choose blog.

Now that the `onboarding-blog` flow is no more we can remove this hack.