Checkout: Remove unsafe lifecycle from CreditCardPaymentBox (#34697)
This replaces the `UNSAFE_componentWillReceiveProps` method with the
supported `componentDidUpdate` in `CreditCardPaymentBox`.

The method is used to start and stop an interval timer which moves the
progress bar along to simulate loading. The loading timer should only be
started or stopped when the props change such that the progress bar is
shown or has an error, so it requires access to both the current
transactionStep data and the previous one. `componentDidUpdate` has
both, just like `UNSAFE_componentWillReceiveProps`.

This depends on #34694 to simplify the function which determines if the
timer should start.