Hide site preview when showing the loading screen
The site preview during signup was always showing if `showSiteMockups`
was true in `steps-pure.js`, even when the loading screen was displaying
at the end of the flow.

This worked for the `onboarding` flow because the last step just
happened to have no preview. However flows like `premium` end on a step
that shows a preview.

This change does change the DOM a little bit (moving the site preview
inside the `.signup__step` container) but it doesn't look like it
affects any layout. And I think the preview belongs in the same ternary
expression that chooses between `<SignupProcessingScreen>` and
`<CurrentComponent>`.