CreditCardPaymentBox: Discard synthetic event for handlePaymentBoxSubmit
The `StripeElementsPaymentBox` has a submit button which generates a
React Synthetic event. It calls `preventDefault()` on the event and then
passes it along as normal. However,
`SecurePaymentForm.handlePaymentBoxSubmit` also calls `preventDefault()`
on the event, and since it's Synthetic we can't still modify it after
it's been passed to another function. This change discards the event to
prevent a warning.