FSE: Handle post content block with no attributes when merging post with template (#34797)
In order to display the template on the editor, the FSE plugin merges the
content with the template during the the_post action.

Previously, the plugin was assuming that any post content block present
in the template will have attributes, but that won't be true after
Automattic/themes#1103, making the str_replace call to don't wrap the
inner blocks in a post content block, since the expected pattern was
not found.

This PR tries to make that merge more abstract by just copying the same
markup present in the template. If there are no attributes, the wrapped post
content block won't have them.