Fix checklist requests in desktop app
Requesting checklist data failed in the desktop app, which would cause
the checklist page to always show the loading state. This is because
requests to the server would ask for the http_envelop, but the desktop
code that handled the response wouldn't unwrap the envelop because the
of the http action params.