Action Panel: Allow figure to be left aligned (#34606)
* Action Panel: Allow figure to be left aligned

As part of the project to overhaul the 'Earn' section, we need a card
format that's similar to the action panel, but has the image on the
lefthand side.

This change adds a prop to the figure component, that can be set to
`left` or `right`, with a default of right to be consistent with the
current behaviour.

* Change alignment class name on the figure component

It now will have a specific `align-left` or `align-right` class name
applied.

* Add new CTA component for use in the panel body

It was felt that the footer component shouldn't be used within the
`ActionPanel` body. Instead we have a generic CTA container component
(with shared CSS properties) and make the footer use that.

We also turn off the namespace class name lint check for the example.