Components: Add a `PlanGate` component to conditionally display content
As part of the project to redesign the 'Earn' section we're developing
various generic components to help with the layout that can be reused in
other areas. This one conditionally renders one or other of its child
components based on whether the currently selected site has a feature on
its plan.

The component takes the feature name as a prop, and renders the first
child if the feature is unavailable. If the feature is present on the
plan, then the second child is rendered.