Feature Flags: Disable earn-relayout in staging and wpcalypso
We've started committing changes to the earn section behind a feature
flag, but as the flag was enabled in staging, it was being shown to
users not in development and essentially broke this section of the
application.

This disables the feature flag by default for now until the work is
ready for testing.