Components: Add a general promo card component (#34607)
* Action Panel: Allow figure to be left aligned

As part of the project to overhaul the 'Earn' section, we need a card
format that's similar to the action panel, but has the image on the
lefthand side.

This change adds a prop to the figure component, that can be set to
`left` or `right`, with a default of right to be consistent with the
current behaviour.

* Change alignment class name on the figure component

It now will have a specific `align-left` or `align-right` class name
applied.

* Components: Add a general promo card component

As part of the project to overhaul the 'Earn' section, we need a card
component that is similar to the card used in the Marketing Tools
section.

This change creates a generalised version of the card component that can
be used in both sections, along with any future sections that follow a
similar pattern.

* Components: Add a primary version of the PromoCard

This adds an `isPrimary` prop to the PromoCard component, which triggers
the styling used in the header section of the design. There are a couple
of bug fixes as well as a change to the header image - updated to be an
SVG.

* PromoCard: Update exports to include prop interfaces

In order to create the `PromoSection` component we need the interfaces
for the prop types of the cards being displayed.

This commit also renames `FC` to `FunctionComponent`.

* PromoCard: Check for no children

When the `PromoCard` is primary we add a prop to the `PromoCardCta` if
present, but if there are no children the map function could throw a
null reference error. So let's check for null!

* Rebased with `ActionPanel` changes and added onClick callback support

* Action Panel: Allow figure to be left aligned

As part of the project to overhaul the 'Earn' section, we need a card
format that's similar to the action panel, but has the image on the
lefthand side.

This change adds a prop to the figure component, that can be set to
`left` or `right`, with a default of right to be consistent with the
current behaviour.

* Change alignment class name on the figure component

It now will have a specific `align-left` or `align-right` class name
applied.

* Fix the responsive styles

By changing the layout to flexbox and setting a limit on the size of the
figure, it broke the styling on screens less than 480px.

* Updating the CTA component to optionally use the plan gate

* Add the option to display different buttons based on plan feature

* Updated based on feedback about TS style

* Tweaked primary card style so the figure is centred on mobile

* Make Learn more translatable. Remove double quotes so learnMoreLink value is used

* Updated to use main ActionPanel component

There was a problem where the ActionPanel styles were not being loaded
and that was because only the sub-components were being used. In order
to use the ActionPanel it needed tweaking to allow extra class names to
be added to the card element.