Me: CreditCardForm: Refactor as functional component (#34963)
* CreditCardForm: Refactor to functional component

* CreditCardForm: Remove upgrades/checkout flag

This flag has been enabled in production since 2015. I doubt it will
ever be needed again.

* CreditCardForm Test: Import getParamsForApi

During refactoring of `CreditCardForm`, `getParamsForApi()` was changed
from an object method to a stand-alone function (since it does not
depend on the component's state anyway). This changes the tests to
import and use that function explicitly.

* lib/checkout: Ensure return type of validatePaymentDetails

`validatePaymentDetails()` should return an object of the form:

```
{ errors: {} }
```

However, if `paymentFieldRules()` returned a falsey value, it instead
returned:

```
{ errors: [] }
```

This change makes sure the former form is always returned.

* CreditCardForm Test: Mock i18n.translate

The form validation directly uses `i18n.translate()` rather than an
injected version, and so the test of that function fails without
providing that function.

* CreditCardForm: [Fix] Merge touchedFormFields

* CreditCardForm: Move non-components to helpers

This should make it easier to see where the component starts and ends.

* CreditCardForm Test: Change getParamsForApi to helpers

* CreditCardForm Tests: Add tests for child props