StoredCard: Make expiry and lastDigits optional
When a StoredCard is created directly on the client from the data
entered into a Stripe Elements form, the card number and expiry are not
available since those fields are inside iframes. This prevents errors
when trying to display such card data. When the data is refreshed from
the server, those fields should be available again.