Add caniuse-lite as a direct dependency.
This library is used by caniuse-api, browserslist,
browserslist-useragent, and others. It contains information on what
browser versions are available and what features they contain.

As such, it should be kept up to date and regularly updated, but it's
been falling behind, since Renovate doesn't appear to pick it up as an
indirect dependency.

This change turns it into a direct dependency for that purpose.
It also updates it to the latest available version.