Prevent duplicate Babel object spread helpers.
With version 7.5.x of Babel, the object spread helper was updated to fix
some issues.

When we upgraded Babel to 7.5.5, it started trying to use the new helper
to perform object spreads. This would have been fine, since the relevant
package (transform-runtime) was part of the upgrade, but Babel
sadly assumes that its version is older, instead of auto-detecting.

This change explicitly indicates which version of the transform-runtime
we're using, fixing the issue. It unfortunately adds extra maintenance
overhead to Babel upgrades, but the Babel authors are considering adding
the aforemention auto-detection, at which point we could remove the
explicit definition.

See https://github.com/babel/babel/issues/10261