Checklist menu hidden for existing - and other ineligible - sites
Not showing the checklist menu for existing sites because sites that
have been running successfully for years have no need for a Checklist
item to be added to the menu.

Also using the `isEligibleForDotcomChecklist()` selector to exclude
other sites that don't have a checklist e.g. sites where the
`design_type` option is `'store'`.

Add unit tests to try and set some of the conditions documented more
clearly.