Full Site Editing: refactor template data insertion
In order to avoid duplicating this file in each theme that we want to add FSE support to,
let's add it here and make sure it's always included via mu-plugins/full-site-editing.php.
This should be the case even when FSE functionality is disabled, to allow themes to populate
their data correctly. On Atomic side, we can achieve something similar via wpcomsh.