Full Site Editing: remove Templates from Calypso's sidebar (#35156)
Removes Full Site Editing templates item from the sidebar so they
can't be accessed directly. Note that we can't do this in FSE plugin side
since we'd have to disable show_in_rest option for this CPT, but that
interferes with some templates functionality that we need in Gutenberg
(e.g. the template previews wouldn't work while editing a full page).