Replace some usage of string refs in stats. (#34772)
This change reimplements some string ref usage in stats and the site
selector using the modern `React.createRef()`.

String refs are deprecated and may be dropped by future versions of
React. They will also need to be replaced if we ever decide to switch to
Preact, either partially or completely.