Me: Use Stripe Elements for "Add Card" form (#34848)
* AddCreditCard: Add stripe context with SetupIntent

* lib/stripe: Add createStripeSetupIntent

* state/stored-cards: Fix linting issues

* CreditCardForm: Validate in Stripe mode

This passes the 'stripe' parameter to `validatePaymentDetails()` when
the stripe form is showing, causing the validation to ignore the stripe
fields (which have their own validation).

* CreateCardForm: Transform createCardToken to async before saveCreditCard

This way `saveCreditCard` can accept any async function that returns a
token. We have to handle stripe tokens a little differently since they
return `id` instead of `token` in the response.

This also adds a flow to use a SetupIntent if stripe is enabled.

* CreditCardForm: Add debug info for submission errors

* UndocumentedMe: Allow additional data in storedCardAdd

This will allow arbitrary data (in this case, `payment_partner`) to be
passed when creating the stored card.

Since we were already modifying the signature of the function, this also
removes the optional callback argument. The function already returns a
Promise.

* CreditCardForm: Simplify saveCreditCard

This just splits saveCreditCard into smaller functions to make it easier
to see how each function operates. It also adds an injected parser
function so that the token creation function can return whatever it
likes.

* StoredCard: Make expiry and lastDigits optional

When a StoredCard is created directly on the client from the data
entered into a Stripe Elements form, the card number and expiry are not
available since those fields are inside iframes. This prevents errors
when trying to display such card data. When the data is refreshed from
the server, those fields should be available again.

* CreditCardForm: Pass isStripeLoading through to CreditCardFormFields