Remove unneeded disables of no-unused-vars ESLint rule
Most of them are not needed since we enabled the `ignoreRestSiblings` flag that allows
usage like this:
```js
const { a, b, c, ...otherProps } = props;
```
which creates an `otherProps` object by removing fields `a`, `b` and `c` from `props`.
The `a`, `b`, `c` variables are not supposed to be used and ESLint knows it.