CreditCardForm: Handle stripe errors specially
If there is a stripe error, often the error code or message will not be
useful to the user. A secondary issue is that if the setup intent is
created, used, and an error happens, the setup intent must be re-created
to submit the form again.

This change handles stripe errors with special logic, displaying a
generic message and causing the setup intent to be recreated.