Update the checklist return URLs for Jetpack sites
This adds a check for the new home section and if enabled ensures that
the Jetpack customers are redirected to that section rather than the
`/my-plans/plan/:site` page.

If the Jetpack connect flow is visited by a non-Jetpack site this will
also redirect to the home section rather than the previous plan page.